requirements.txt
Django==1.11.6
django-rest-framework==0.1.0
django-templated-mail==1.0.0
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
djoser==1.0.1
olefile==0.44
Pillow==4.3.0
PyJWT==1.5.3
pytz==2017.2
